Refactor: 패키지들을 추가했다
추가된 패키지명: controller, input, util, view, model
