+++ src/main/java/domain/model/card/Type.java
@@ -1,4 +1,4 @@
-package domain.card;
+package domain.model.card;
 
 public enum Type {
     SPADE,
