+++ src/test/java/domain/model/card/CardFactoryTest.java
@@ -1,4 +1,4 @@
-package domain.card;
+package domain.model.card;
 
 import org.junit.jupiter.api.Test;
